core/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
